pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "objectwire"
version = "0.1.0"
description = "AI-powered RSS/URL scraper agent for blockchain prediction markets"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "feedparser>=6.0.0",
]

[project.scripts]
objectwire = "objectwire.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/objectwire"]
